fix: creates Conference Key the proper way
from Udacity's review: This is not the correct way to create a Conference
Key and as a result the ancestral relation will not be properly
initialized. The reason is that the key to the actual Conference is a
child from a Profile object, so building a Key from just its id is
insufficient. To convince yourself of this, go in to the admin viewer
after creating a Session and click on its 'ancestor' link. You'll find
there is no such entity!
To correct this bug, simply use the proper Conference Key like you did on
line 664 (without the .get()) or simply use that Conference entity's key:
conference.key.